refactor(adapters): split provider_docker into package modules
- Extract types, images, package_view, multi_actor, and provider orchestration
- Update L1 architecture boundary scan for the package layout (Refs: #31)
