+++ src/bora/adapters/provider_docker/__init__.py
@@ -1,0 +1,27 @@
+"""Docker L1 Provider package (chore #31 split of provider_docker monolith).
+
+Layout: types · errors · package_view · images · multi_actor · provider.
+"""
+
+from __future__ import annotations
+
+from bora.adapters.provider_docker.errors import ProviderL1Error
+from bora.adapters.provider_docker.images import (
+    build_package_image,
+    ensure_base_image,
+    ensure_image_lock,
+    sys_executable,
+)
+from bora.adapters.provider_docker.provider import DockerProvider
+from bora.adapters.provider_docker.types import DockerImageLock, DockerRuntime
+
+__all__ = [
+    "DockerImageLock",
+    "DockerProvider",
+    "DockerRuntime",
+    "ProviderL1Error",
+    "build_package_image",
+    "ensure_base_image",
+    "ensure_image_lock",
+    "sys_executable",
+]
