+++ src/bora/adapters/provider_docker/errors.py
@@ -1,0 +1,9 @@
+"""L1 Provider error types."""
+
+from __future__ import annotations
+
+from bora.provider.errors import ProviderError
+
+
+class ProviderL1Error(ProviderError):
+    """L1-specific errors with stable kinds from Spec 07."""
