+++ tests/architecture/test_provider_l1_boundaries.py
@@ -15,6 +15,7 @@ def test_cli_does_not_import_docker_sdk() -> None:
 
 
 def test_provider_docker_no_benchmark_branches() -> None:
-    text = (ROOT / "src" / "bora" / "adapters" / "provider_docker.py").read_text(encoding="utf-8")
+    pkg = ROOT / "src" / "bora" / "adapters" / "provider_docker"
+    text = "\n".join(p.read_text(encoding="utf-8") for p in sorted(pkg.rglob("*.py")))
     for banned in ("terminal-bench", "TerminalBench", "multiagentbench", "tau2"):
         assert banned not in text
